+++ BraintreeCore/src/main/java/com/braintreepayments/api/core/AnalyticsApi.kt
@@ -89,6 +89,7 @@ internal class AnalyticsApi(
             .putOpt(FPTI_KEY_ERROR_DESC, event.errorDescription)
             .putOpt(FPTI_KEY_CONTEXT_TYPE, if (event.isVaultRequest) "BA-TOKEN" else "EC-TOKEN")
             .putOpt(FPTI_KEY_PAYPAL_ATTEMPTED_APP_SWITCH, event.didSdkAttemptAppSwitch)
+            .putOpt(FPTI_KEY_FUNDING_SOURCE, event.fundingSource)
     }
 
     @Throws(JSONException::class)
@@ -146,6 +147,7 @@ internal class AnalyticsApi(
         private const val FPTI_KEY_PAYPAL_ATTEMPTED_APP_SWITCH = "attempted_app_switch"
         private const val FPTI_KEY_ERROR_DESC = "error_desc"
         private const val FPTI_KEY_CONTEXT_TYPE = "context_type"
+        private const val FPTI_KEY_FUNDING_SOURCE = "funding_source"
 
         private const val FPTI_BATCH_KEY_VENMO_INSTALLED = "venmo_installed"
         private const val FPTI_BATCH_KEY_PAYPAL_INSTALLED = "paypal_installed"