+++ BraintreeCore/src/main/java/com/braintreepayments/api/core/AnalyticsParamRepository.kt
@@ -32,6 +32,11 @@ class AnalyticsParamRepository(
     var didSdkAttemptAppSwitch: Boolean? = null
 
     /**
+     * Value to determine whether PayPal, PayLater or PayPal Credit was clicked.
+     */
+    var fundingSource: String? = null
+
+    /**
      * Session ID to tie analytics events together which is used for reporting conversion funnels.
      */
     val sessionId: String = uuidHelper.formattedUUID
@@ -48,6 +53,7 @@ class AnalyticsParamRepository(
         didEnablePayPalAppSwitch = null
         didPayPalServerAttemptAppSwitch = null
         didSdkAttemptAppSwitch = null
+        fundingSource = null
     }
 
     companion object {